+++ node/toolrunner.ts
@@ -2,7 +2,6 @@ import Q = require('q');
 import os = require('os');
 import events = require('events');
 import child = require('child_process');
-import stream = require('stream');
 import im = require('./internal');
 import fs = require('fs');
 
@@ -98,7 +97,7 @@ export class ToolRunner extends events.EventEmitter {
                 if (c !== '"') {
                     arg += '\\';
                 } else {
-                    arg.slice(0, -1);         
+                    arg.slice(0, -1);
                 }
             }
             arg += c;
@@ -174,7 +173,7 @@ export class ToolRunner extends events.EventEmitter {
             // Windows (regular)
             else {
                 commandParts.push(this._windowsQuoteCmdArg(toolPath));
-                commandParts = commandParts.concat(args.map(arg =>this._windowsQuoteCmdArg(arg)));
+                commandParts = commandParts.concat(args.map(arg => this._windowsQuoteCmdArg(arg)));
             }
         }
         else {
@@ -194,27 +193,27 @@ export class ToolRunner extends events.EventEmitter {
         return cmd;
     }
 
-    private _processLineBuffer(data: Buffer, strBuffer: string, onLine: (line: string) => void): void {
-        try {
-            var s = strBuffer + data.toString();
-            var n = s.indexOf(os.EOL);
+    private _processLineBuffer(data: Buffer, buffer: string, onLine: (line: string) => void): string {
+        let newBuffer = buffer + data.toString();
 
-            while (n > -1) {
-                var line = s.substring(0, n);
+        try {
+            let eolIndex = newBuffer.indexOf(os.EOL);
+
+            while (eolIndex > -1) {
+                const line = newBuffer.substring(0, eolIndex);
                 onLine(line);
 
                 // the rest of the string ...
-                s = s.substring(n + os.EOL.length);
-                n = s.indexOf(os.EOL);
+                newBuffer = newBuffer.substring(eolIndex + os.EOL.length);
+                eolIndex = newBuffer.indexOf(os.EOL);
             }
-
-            strBuffer = s;
         }
         catch (err) {
             // streaming lines to console is best effort.  Don't fail a build.
             this._debug('error processing line');
         }
 
+        return newBuffer;
     }
 
     /**
@@ -333,7 +332,7 @@ export class ToolRunner extends events.EventEmitter {
                 return args;
             }
         } else if (options.shell) {
-            return this.args.map(arg =>  {
+            return this.args.map(arg => {
                 if (this._isWrapped(arg, "'")) {
                     return arg;
                 }
@@ -636,7 +635,7 @@ export class ToolRunner extends events.EventEmitter {
             this._getSpawnArgs(optionsNonNull),
             this._getSpawnOptions(optionsNonNull));
 
-        waitingEvents ++;
+        waitingEvents++;
         cp = child.spawn(
             pipeOutputToTool._getSpawnFileName(optionsNonNull),
             pipeOutputToTool._getSpawnArgs(optionsNonNull),
@@ -650,7 +649,7 @@ export class ToolRunner extends events.EventEmitter {
                 fileStream.on('finish', () => {
                     waitingEvents--; //file write is complete
                     fileStream = null;
-                    if(waitingEvents == 0) {
+                    if (waitingEvents == 0) {
                         if (error) {
                             reject(error);
                         } else {
@@ -662,7 +661,7 @@ export class ToolRunner extends events.EventEmitter {
                     waitingEvents--; //there were errors writing to the file, write is done
                     this._debug(`Failed to pipe output of ${toolPathFirst} to file ${this.pipeOutputToFile}. Error = ${err}`);
                     fileStream = null;
-                    if(waitingEvents == 0) {
+                    if (waitingEvents == 0) {
                         if (error) {
                             reject(error);
                         } else {
@@ -671,7 +670,7 @@ export class ToolRunner extends events.EventEmitter {
                     }
                 });
             }
-    
+
             //pipe stdout of first tool to stdin of second tool
             cpFirst.stdout?.on('data', (data: Buffer) => {
                 try {
@@ -703,7 +702,7 @@ export class ToolRunner extends events.EventEmitter {
                 }
                 cp.stdin?.end();
                 error = new Error(toolPathFirst + ' failed. ' + err.message);
-                if(waitingEvents == 0) {
+                if (waitingEvents == 0) {
                     reject(error);
                 }
             });
@@ -719,7 +718,7 @@ export class ToolRunner extends events.EventEmitter {
                     fileStream.end();
                 }
                 cp.stdin?.end();
-                if(waitingEvents == 0) {
+                if (waitingEvents == 0) {
                     if (error) {
                         reject(error);
                     } else {
@@ -727,69 +726,69 @@ export class ToolRunner extends events.EventEmitter {
                     }
                 }
             });
-    
-            var stdbuffer: string = '';
+
+            let stdLineBuffer = '';
             cp.stdout?.on('data', (data: Buffer) => {
                 this.emit('stdout', data);
-    
+
                 if (!optionsNonNull.silent) {
                     optionsNonNull.outStream!.write(data);
                 }
-    
-                this._processLineBuffer(data, stdbuffer, (line: string) => {
+
+                stdLineBuffer = this._processLineBuffer(data, stdLineBuffer, (line: string) => {
                     this.emit('stdline', line);
                 });
             });
-    
-            var errbuffer: string = '';
+
+            let errLineBuffer = '';
             cp.stderr?.on('data', (data: Buffer) => {
                 this.emit('stderr', data);
-    
+
                 success = !optionsNonNull.failOnStdErr;
                 if (!optionsNonNull.silent) {
                     var s = optionsNonNull.failOnStdErr ? optionsNonNull.errStream! : optionsNonNull.outStream!;
                     s.write(data);
                 }
-    
-                this._processLineBuffer(data, errbuffer, (line: string) => {
+
+                errLineBuffer = this._processLineBuffer(data, errLineBuffer, (line: string) => {
                     this.emit('errline', line);
                 });
             });
-    
+
             cp.on('error', (err: Error) => {
                 waitingEvents--; //process is done with errors
                 error = new Error(toolPath + ' failed. ' + err.message);
-                if(waitingEvents == 0) {
+                if (waitingEvents == 0) {
                     reject(error);
                 }
             });
-    
+
             cp.on('close', (code: number, signal: any) => {
                 waitingEvents--; //process is complete
                 this._debug('rc:' + code);
                 returnCode = code;
-    
-                if (stdbuffer.length > 0) {
-                    this.emit('stdline', stdbuffer);
+
+                if (stdLineBuffer.length > 0) {
+                    this.emit('stdline', stdLineBuffer);
                 }
-    
-                if (errbuffer.length > 0) {
-                    this.emit('errline', errbuffer);
+
+                if (errLineBuffer.length > 0) {
+                    this.emit('errline', errLineBuffer);
                 }
-    
+
                 if (code != 0 && !optionsNonNull.ignoreReturnCode) {
                     success = false;
                 }
-    
+
                 this._debug('success:' + success);
-    
+
                 if (!successFirst) { //in the case output is piped to another tool, check exit code of both tools
                     error = new Error(toolPathFirst + ' failed with return code: ' + returnCodeFirst);
                 } else if (!success) {
                     error = new Error(toolPath + ' failed with return code: ' + code);
                 }
-    
-                if(waitingEvents == 0) {
+
+                if (waitingEvents == 0) {
                     if (error) {
                         reject(error);
                     } else {
@@ -838,7 +837,7 @@ export class ToolRunner extends events.EventEmitter {
             this._getSpawnArgs(optionsNonNull),
             this._getSpawnOptions(optionsNonNull));
 
-        waitingEvents ++;
+        waitingEvents++;
         cp = child.spawn(
             pipeOutputToTool._getSpawnFileName(optionsNonNull),
             pipeOutputToTool._getSpawnArgs(optionsNonNull),
@@ -850,7 +849,7 @@ export class ToolRunner extends events.EventEmitter {
             fileStream.on('finish', () => {
                 waitingEvents--; //file write is complete
                 fileStream = null;
-                if(waitingEvents == 0) {
+                if (waitingEvents == 0) {
                     if (error) {
                         defer.reject(error);
                     } else {
@@ -862,7 +861,7 @@ export class ToolRunner extends events.EventEmitter {
                 waitingEvents--; //there were errors writing to the file, write is done
                 this._debug(`Failed to pipe output of ${toolPathFirst} to file ${this.pipeOutputToFile}. Error = ${err}`);
                 fileStream = null;
-                if(waitingEvents == 0) {
+                if (waitingEvents == 0) {
                     if (error) {
                         defer.reject(error);
                     } else {
@@ -901,7 +900,7 @@ export class ToolRunner extends events.EventEmitter {
             }
             cp.stdin?.end();
             error = new Error(toolPathFirst + ' failed. ' + err.message);
-            if(waitingEvents == 0) {
+            if (waitingEvents == 0) {
                 defer.reject(error);
             }
         });
@@ -917,7 +916,7 @@ export class ToolRunner extends events.EventEmitter {
                 fileStream.end();
             }
             cp.stdin?.end();
-            if(waitingEvents == 0) {
+            if (waitingEvents == 0) {
                 if (error) {
                     defer.reject(error);
                 } else {
@@ -926,7 +925,7 @@ export class ToolRunner extends events.EventEmitter {
             }
         });
 
-        var stdbuffer: string = '';
+        let stdLineBuffer = '';
         cp.stdout?.on('data', (data: Buffer) => {
             this.emit('stdout', data);
 
@@ -934,12 +933,12 @@ export class ToolRunner extends events.EventEmitter {
                 optionsNonNull.outStream!.write(data);
             }
 
-            this._processLineBuffer(data, stdbuffer, (line: string) => {
+            stdLineBuffer = this._processLineBuffer(data, stdLineBuffer, (line: string) => {
                 this.emit('stdline', line);
             });
         });
 
-        var errbuffer: string = '';
+        let errLineBuffer = '';
         cp.stderr?.on('data', (data: Buffer) => {
             this.emit('stderr', data);
 
@@ -949,7 +948,7 @@ export class ToolRunner extends events.EventEmitter {
                 s.write(data);
             }
 
-            this._processLineBuffer(data, errbuffer, (line: string) => {
+            errLineBuffer = this._processLineBuffer(data, errLineBuffer, (line: string) => {
                 this.emit('errline', line);
             });
         });
@@ -957,7 +956,7 @@ export class ToolRunner extends events.EventEmitter {
         cp.on('error', (err: Error) => {
             waitingEvents--; //process is done with errors
             error = new Error(toolPath + ' failed. ' + err.message);
-            if(waitingEvents == 0) {
+            if (waitingEvents == 0) {
                 defer.reject(error);
             }
         });
@@ -967,12 +966,12 @@ export class ToolRunner extends events.EventEmitter {
             this._debug('rc:' + code);
             returnCode = code;
 
-            if (stdbuffer.length > 0) {
-                this.emit('stdline', stdbuffer);
+            if (stdLineBuffer.length > 0) {
+                this.emit('stdline', stdLineBuffer);
             }
 
-            if (errbuffer.length > 0) {
-                this.emit('errline', errbuffer);
+            if (errLineBuffer.length > 0) {
+                this.emit('errline', errLineBuffer);
             }
 
             if (code != 0 && !optionsNonNull.ignoreReturnCode) {
@@ -987,7 +986,7 @@ export class ToolRunner extends events.EventEmitter {
                 error = new Error(toolPath + ' failed with return code: ' + code);
             }
 
-            if(waitingEvents == 0) {
+            if (waitingEvents == 0) {
                 if (error) {
                     defer.reject(error);
                 } else {
@@ -1100,22 +1099,25 @@ export class ToolRunner extends events.EventEmitter {
             this._debug(message);
         });
 
-        var stdbuffer: string = '';
-        var errbuffer: string = '';
-        const emitDoneEvent = function (resolve, reject) {
+        let stdLineBuffer = '';
+        let errLineBuffer = '';
+        const emitDoneEvent = (
+            resolve: (code: number) => void,
+            reject: (error: Error) => void
+        ) => {
             state.on('done', (error: Error, exitCode: number) => {
-                if (stdbuffer.length > 0) {
-                    this.emit('stdline', stdbuffer);
+                if (stdLineBuffer.length > 0) {
+                    this.emit('stdline', stdLineBuffer);
                 }
-    
-                if (errbuffer.length > 0) {
-                    this.emit('errline', errbuffer);
+
+                if (errLineBuffer.length > 0) {
+                    this.emit('errline', errLineBuffer);
                 }
-    
+
                 if (cp) {
                     cp.removeAllListeners();
                 }
-                
+
                 if (error) {
                     reject(error);
                 }
@@ -1126,7 +1128,7 @@ export class ToolRunner extends events.EventEmitter {
         }
 
         // Edge case when the node itself cant's spawn and emit event
-        let cp;
+        let cp: child.ChildProcess;
         try {
             cp = child.spawn(this._getSpawnFileName(options), this._getSpawnArgs(optionsNonNull), this._getSpawnOptions(options));
         } catch (error) {
@@ -1156,7 +1158,7 @@ export class ToolRunner extends events.EventEmitter {
                 optionsNonNull.outStream!.write(data);
             }
 
-            this._processLineBuffer(data, stdbuffer, (line: string) => {
+            stdLineBuffer = this._processLineBuffer(data, stdLineBuffer, (line: string) => {
                 this.emit('stdline', line);
             });
         });
@@ -1170,7 +1172,7 @@ export class ToolRunner extends events.EventEmitter {
                 s.write(data);
             }
 
-            this._processLineBuffer(data, errbuffer, (line: string) => {
+            errLineBuffer = this._processLineBuffer(data, errLineBuffer, (line: string) => {
                 this.emit('errline', line);
             });
         });
@@ -1182,18 +1184,19 @@ export class ToolRunner extends events.EventEmitter {
             state.CheckComplete();
         });
 
-        cp.on('exit', (code: number, signal: number | NodeJS.Signals) => {
+        // Do not write debug logs here. Sometimes stdio not closed yet and you can damage user output commands.
+        cp.on('exit', (code: number | null, signal: NodeJS.Signals | null) => {
             state.processExitCode = code;
+            state.processExitSignal = signal;
             state.processExited = true;
-            this._debug(`STDIO streams have closed and received exit code ${code} and signal ${signal} for tool '${this.toolPath}'`);
             state.CheckComplete()
         });
 
-        cp.on('close', (code: number, signal: number | NodeJS.Signals) => {
-            state.processExitCode = code;
-            state.processExited = true;
+        cp.on('close', (code: number | null, signal: NodeJS.Signals | null) => {
+            state.processCloseCode = code;
+            state.processCloseSignal = signal;
             state.processClosed = true;
-            this._debug(`STDIO streams have closed and received exit code ${code} and signal ${signal} for tool '${this.toolPath}'`);
+            state.processExited = true;
             state.CheckComplete();
         });
 
@@ -1233,15 +1236,15 @@ export class ToolRunner extends events.EventEmitter {
             this._debug(message);
         });
 
-        var stdbuffer: string = '';
-        var errbuffer: string = '';
+        let stdLineBuffer = '';
+        let errLineBuffer = '';
         state.on('done', (error: Error, exitCode: number) => {
-            if (stdbuffer.length > 0) {
-                this.emit('stdline', stdbuffer);
+            if (stdLineBuffer.length > 0) {
+                this.emit('stdline', stdLineBuffer);
             }
 
-            if (errbuffer.length > 0) {
-                this.emit('errline', errbuffer);
+            if (errLineBuffer.length > 0) {
+                this.emit('errline', errLineBuffer);
             }
 
             if (cp) {
@@ -1258,7 +1261,7 @@ export class ToolRunner extends events.EventEmitter {
 
 
         // Edge case when the node itself cant's spawn and emit event
-        let cp;
+        let cp: child.ChildProcess;
         try {
             cp = child.spawn(this._getSpawnFileName(options), this._getSpawnArgs(optionsNonNull), this._getSpawnOptions(options));
         } catch (error) {
@@ -1287,7 +1290,7 @@ export class ToolRunner extends events.EventEmitter {
                 optionsNonNull.outStream!.write(data);
             }
 
-            this._processLineBuffer(data, stdbuffer, (line: string) => {
+            stdLineBuffer = this._processLineBuffer(data, stdLineBuffer, (line: string) => {
                 this.emit('stdline', line);
             });
         });
@@ -1302,7 +1305,7 @@ export class ToolRunner extends events.EventEmitter {
                 s.write(data);
             }
 
-            this._processLineBuffer(data, errbuffer, (line: string) => {
+            errLineBuffer = this._processLineBuffer(data, errLineBuffer, (line: string) => {
                 this.emit('errline', line);
             });
         });
@@ -1314,18 +1317,19 @@ export class ToolRunner extends events.EventEmitter {
             state.CheckComplete();
         });
 
-        cp.on('exit', (code: number, signal: number | NodeJS.Signals) => {
+        // Do not write debug logs here. Sometimes stdio not closed yet and you can damage user output commands.
+        cp.on('exit', (code: number | null, signal: NodeJS.Signals | null) => {
             state.processExitCode = code;
+            state.processExitSignal = signal;
             state.processExited = true;
-            this._debug(`STDIO streams have closed and received exit code ${code} and signal ${signal} for tool '${this.toolPath}'`);
             state.CheckComplete()
         });
 
-        cp.on('close', (code: number, signal: number | NodeJS.Signals) => {
-            state.processExitCode = code;
-            state.processExited = true;
+        cp.on('close', (code: number | null, signal: NodeJS.Signals | null) => {
+            state.processCloseCode = code;
+            state.processCloseSignal = signal;
             state.processClosed = true;
-            this._debug(`STDIO streams have closed and received exit code ${code} and signal ${signal} for tool '${this.toolPath}'`);
+            state.processExited = true;
             state.CheckComplete();
         });
 
@@ -1404,15 +1408,22 @@ class ExecState extends events.EventEmitter {
     }
 
     processClosed: boolean; // tracks whether the process has exited and stdio is closed
-    processError: string;
-    processExitCode: number;
+    processCloseCode: number | null;
+    processCloseSignal: NodeJS.Signals | null;
+
     processExited: boolean; // tracks whether the process has exited
+    processExitCode: number | null;
+    processExitSignal: NodeJS.Signals | null;
+
+    processError: string;
     processStderr: boolean; // tracks whether stderr was written to
-    private delay = 10000; // 10 seconds
-    private done: boolean;
-    private options: IExecOptions;
+
+    private readonly delay: number = 10000; // 10 seconds
+    private readonly options: IExecOptions;
+    private readonly toolPath: string;
+
     private timeout: NodeJS.Timer | null = null;
-    private toolPath: string;
+    private done: boolean;
 
     public CheckComplete(): void {
         if (this.done) {
@@ -1435,6 +1446,8 @@ class ExecState extends events.EventEmitter {
         // determine whether there is an error
         let error: Error | undefined;
         if (this.processExited) {
+            this._debug(`Process exited with code ${this.processExitCode} and signal ${this.processExitSignal} for tool '${this.toolPath}'`);
+
             if (this.processError) {
                 error = new Error(im._loc('LIB_ProcessError', this.toolPath, this.processError));
             }
@@ -1444,6 +1457,10 @@ class ExecState extends events.EventEmitter {
             else if (this.processStderr && this.options.failOnStdErr) {
                 error = new Error(im._loc('LIB_ProcessStderr', this.toolPath));
             }
+        }
+
+        if (this.processClosed) {
+            this._debug(`STDIO streams have closed and received exit code ${this.processCloseCode} and signal ${this.processCloseSignal} for tool '${this.toolPath}'`);
         }
 
         // clear the timeout