+++ node/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## 4.x
 
+## 4.17.2
+
+- Fix ToolRunner stdline/errline events buffering - [#1055](https://github.com/microsoft/azure-pipelines-task-lib/pull/1055)
+
+## 4.17.1
+
+- Fix debug logs inside user commands - [#1064](https://github.com/microsoft/azure-pipelines-task-lib/pull/1064)
+
 ## 4.17.0
 
 - Added signal handler for process execution to kill process with proper signal - [#1008](https://github.com/microsoft/azure-pipelines-task-lib/pull/1008)